Ensure idempotent installation of Google Cloud SDK

+++ pkgs/install.sh
@@ -35,8 +35,10 @@ else
 
   # Install Google Cloud SDK
   echo "*** Installing Google Cloud SDK..."
-  curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo gpg --dearmor -o /usr/share/keyrings/cloud.google.gpg
-  echo "deb [signed-by=/usr/share/keyrings/cloud.google.gpg] https://packages.cloud.google.com/apt cloud-sdk main" | sudo tee -a /etc/apt/sources.list.d/google-cloud-sdk.list
+  if [ ! -f /usr/share/keyrings/cloud.google.gpg ]; then
+    curl https://packages.cloud.google.com/apt/doc/apt-key.gpg | sudo gpg --dearmor -o /usr/share/keyrings/cloud.google.gpg
+    echo "deb [signed-by=/usr/share/keyrings/cloud.google.gpg] https://packages.cloud.google.com/apt cloud-sdk main" | sudo tee -a /etc/apt/sources.list.d/google-cloud-sdk.list
+  fi
   sudo apt-get update
   sudo apt-get -y install google-cloud-cli
   sudo apt-get -y install google-cloud-sdk-gke-gcloud-auth-plugin
